+++ tests/profiles/name_test.py
@@ -10,6 +10,7 @@ from fontbakery.constants import (NameID,
                                   MacintoshLanguageID)
 from fontbakery.checkrunner import INFO, WARN, PASS, FAIL, SKIP
 from fontbakery.codetesting import (assert_PASS,
+                                    assert_SKIP,
                                     assert_results_contain,
                                     CheckTester,
                                     portable_path,
@@ -493,3 +494,68 @@ def test_check_family_max_4_fonts_per_family_name():
     assert_results_contain(check(test_fonts),
                            FAIL, 'too-many')
 
+
+def test_check_italic_names():
+    check = CheckTester(opentype_profile,
+                        "com.google.fonts/check/name/italic_names")
+
+    def get_name(font, nameID):
+        for entry in font['name'].names:
+            if entry.nameID == nameID:
+                return entry.toUnicode()
+
+    def set_name(font, nameID, string):
+        for record in font["name"].names:
+            if record.nameID == nameID:
+                old_string = record.toUnicode()
+                if string != old_string:
+                    font["name"].setName(
+                        string,
+                        record.nameID,
+                        record.platformID,
+                        record.platEncID,
+                        record.langID,
+                    )
+
+    # Fonts without Name ID 16 & 17
+
+    # PASS or SKIP
+    ttFont = TTFont(TEST_FILE("cabin/Cabin-Regular.ttf"))
+    assert_SKIP(check(ttFont))
+
+    ttFont = TTFont(TEST_FILE("cabin/Cabin-Italic.ttf"))
+    assert_PASS(check(ttFont), PASS)
+
+    ttFont = TTFont(TEST_FILE("cabin/Cabin-Bold.ttf"))
+    assert_SKIP(check(ttFont))
+
+    ttFont = TTFont(TEST_FILE("cabin/Cabin-BoldItalic.ttf"))
+    assert_PASS(check(ttFont), PASS)
+
+    # FAIL
+    ttFont = TTFont(TEST_FILE("cabin/Cabin-Italic.ttf"))
+    set_name(ttFont, 1, get_name(ttFont, 1) + " Italic")
+    assert_results_contain(check(ttFont), FAIL, "bad-familyname")
+
+    ttFont = TTFont(TEST_FILE("cabin/Cabin-Italic.ttf"))
+    set_name(ttFont, 2, "Regular")
+    assert_results_contain(check(ttFont), FAIL, "bad-subfamilyname")
+
+
+    # Fonts with Name ID 16 & 17
+
+    # PASS or SKIP
+    ttFont = TTFont(TEST_FILE("shantell/ShantellSans[BNCE,INFM,SPAC,wght].ttf"))
+    assert_SKIP(check(ttFont))
+
+    ttFont = TTFont(TEST_FILE("shantell/ShantellSans-Italic[BNCE,INFM,SPAC,wght].fixed.ttf"))
+    assert_PASS(check(ttFont), PASS)
+
+    # FAIL
+    ttFont = TTFont(TEST_FILE("shantell/ShantellSans-Italic[BNCE,INFM,SPAC,wght].fixed.ttf"))
+    set_name(ttFont, 16, "Shantell Sans Italic")
+    assert_results_contain(check(ttFont), FAIL, "bad-typographicfamilyname")
+
+    ttFont = TTFont(TEST_FILE("shantell/ShantellSans-Italic[BNCE,INFM,SPAC,wght].fixed.ttf"))
+    set_name(ttFont, 17, "Light")
+    assert_results_contain(check(ttFont), FAIL, "bad-typographicsubfamilyname")